backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.9
scikit-learn==1.4.1.post1
numpy==1.26.4
pandas==2.2.1
paho-mqtt==1.6.1
websockets==12.0
python-dotenv==1.0.1
